Add fixes for audio not playing correctly
This adds a check to AudioContext to only load once user presses on the page.

It also adds a check for first load of the page to not show the loading symbol when autoplay doesn't work.

This has an issues on browsers that do autoplay, as the stream will load anyway. If the user clicks play, it will load the stream again.

This has been decided as a good compromise.